+++ projects/ngx-tree-dnd-fork/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/ngx-tree-dnd",
+  "dest": "../../dist/ngx-tree-dnd-fork",
   "deleteDestPath": false,
   "lib": {
     "entryFile": "src/public_api.ts"
