+++ projects/ngx-tree-dnd-fork/src/lib/models/tree-view.enum.ts
