+++ projects/ngx-tree-dnd-fork/src/lib/ngx-tree-dnd-fork.module.ts
@@ -3,11 +3,11 @@ import { NgModule } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-import { AutoFocusDirective } from './directives/ngx-tree-dnd-autofocus.directive';
-import { DragElementsDirective } from './directives/ngx-tree-dnd-drag.directive';
-import { DropElementsDirective } from './directives/ngx-tree-dnd-drop.directive';
-import { NgxTreeParentComponent } from './ngx-tree-dnd-parent/ngx-tree-dnd-parent.component';
-import { NgxTreeChildrenComponent } from './ngx-tree-dnd-children/ngx-tree-dnd-children.component';
+import { AutoFocusDirective } from './directives/ngx-tree-dnd-fork-autofocus.directive';
+import { DragElementsDirective } from './directives/ngx-tree-dnd-fork-drag.directive';
+import { DropElementsDirective } from './directives/ngx-tree-dnd-fork-drop.directive';
+import { NgxTreeParentComponent } from './ngx-tree-dnd-fork-parent/ngx-tree-dnd-fork-parent.component';
+import { NgxTreeChildrenComponent } from './ngx-tree-dnd-fork-children/ngx-tree-dnd-fork-children.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faCoffee, faPlus, faEdit, faMinus, faTimes, faCheck, faArrowDown } from '@fortawesome/free-solid-svg-icons';
