+++ projects/ngx-tree-dnd-fork/src/lib/styles-tree-dnd-fork.css
@@ -1,7 +1,7 @@
 /*
  Copyright (C) 2018 Yaroslav Kikot
  This project is licensed under the terms of the MIT license.
- https://github.com/Zicrael/ngx-tree-dnd
+ https://github.com/Zicrael/ngx-tree-dnd-fork
  */
 
 /* Tree main styles */
