--- projects/ngx-tree-dnd/src/public_api.ts
@@ -1,12 +1,0 @@
-/*
- * Public API Surface of ngx-tree-dnd
- */
-export * from './lib/ngx-tree-dnd.service';
-export * from './lib/ngx-tree-dnd-parent/ngx-tree-dnd-parent.component';
-export * from './lib/ngx-tree-dnd-children/ngx-tree-dnd-children.component';
-export * from './lib/directives/ngx-tree-dnd-autofocus.directive';
-export * from './lib/directives/ngx-tree-dnd-drag.directive';
-export * from './lib/directives/ngx-tree-dnd-drop.directive';
-export * from './lib/models/tree-view.model';
-export * from './lib/models/tree-view.enum';
-export * from './lib/ngx-tree-dnd.module';
